Moved most class declarations from headers to their on cc file
