+++ src/UiDateTimePicker.cc
@@ -2,6 +2,12 @@
 #include "ui-node.h"
 #include "nbind/api.h"
 
+class UiDateTimePicker : public UiControl {
+  public:
+	UiDateTimePicker();
+	DEFINE_CONTROL_METHODS()
+};
+
 UiDateTimePicker::UiDateTimePicker()
 	: UiControl((uiControl *)uiNewDateTimePicker()) {}
 
@@ -12,6 +18,12 @@ NBIND_CLASS(UiDateTimePicker) {
 	DECLARE_CHILD_CONTROL_METHODS()
 }
 
+class UiTimePicker : public UiControl {
+  public:
+	UiTimePicker();
+	DEFINE_CONTROL_METHODS()
+};
+
 UiTimePicker::UiTimePicker() : UiControl((uiControl *)uiNewTimePicker()) {}
 
 INHERITS_CONTROL_METHODS(UiTimePicker)
@@ -20,6 +32,12 @@ NBIND_CLASS(UiTimePicker) {
 	construct<>();
 	DECLARE_CHILD_CONTROL_METHODS()
 }
+
+class UiDatePicker : public UiControl {
+  public:
+	UiDatePicker();
+	DEFINE_CONTROL_METHODS()
+};
 
 UiDatePicker::UiDatePicker() : UiControl((uiControl *)uiNewDatePicker()) {}
