+++ src/UiGroup.cc
@@ -2,6 +2,18 @@
 #include "ui-node.h"
 #include "nbind/api.h"
 
+class UiGroup : public UiControl {
+  public:
+	UiGroup(const char *text);
+	UiGroup();
+	void setChild(UiControl *control);
+	bool getMargined();
+	void setMargined(bool margined);
+	const char *getTitle();
+	void setTitle(const char *title);
+	DEFINE_CONTROL_METHODS()
+};
+
 UiGroup::UiGroup(const char *text) : UiControl((uiControl *)uiNewGroup(text)) {}
 UiGroup::UiGroup() : UiControl((uiControl *)uiNewGroup("")) {}
