+++ src/UiTab.cc
@@ -1,33 +1,47 @@
+#include <string>
 #include "../ui.h"
 #include "ui-node.h"
 #include "nbind/api.h"
+
+class UiTab : public UiControl {
+  public:
+	UiTab();
+	void append(std::string text, UiControl *child);
+	void insertAt(std::string name, int before, UiControl *child);
+	void deleteAt(int index);
+	int numPages();
+	bool getMargined(int page);
+	void setMargined(int page, bool margined);
+
+	DEFINE_CONTROL_METHODS()
+};
 
 UiTab::UiTab() : UiControl((uiControl *)uiNewTab()) {}
 
 INHERITS_CONTROL_METHODS(UiTab)
 
-void UiTab::append(const char *text, UiControl *child) {
-	uiTabAppend((uiTab *)getHandle(), text, child->getHandle());
+void UiTab::append(std::string text, UiControl *child) {
+	uiTabAppend(uiTab(getHandle()), text.c_str(), child->getHandle());
 }
 
-void UiTab::insertAt(const char *name, int before, UiControl *child) {
-	uiTabInsertAt((uiTab *)getHandle(), name, before, child->getHandle());
+void UiTab::insertAt(std::string name, int before, UiControl *child) {
+	uiTabInsertAt(uiTab(getHandle()), name.c_str(), before, child->getHandle());
 }
 
 void UiTab::deleteAt(int index) {
-	uiTabDelete((uiTab *)getHandle(), index);
+	uiTabDelete(uiTab(getHandle()), index);
 }
 
 int UiTab::numPages() {
-	return uiTabNumPages((uiTab *)getHandle());
+	return uiTabNumPages(uiTab(getHandle()));
 }
 
 bool UiTab::getMargined(int page) {
-	return uiTabMargined((uiTab *)getHandle(), page);
+	return uiTabMargined(uiTab(getHandle()), page);
 }
 
 void UiTab::setMargined(int page, bool margined) {
-	uiTabSetMargined((uiTab *)getHandle(), page, margined);
+	uiTabSetMargined(uiTab(getHandle()), page, margined);
 }
 
 NBIND_CLASS(UiTab) {
